--- lib/services/ios_playback_state_sync.dart
@@ -1,29 +1,0 @@
-import 'dart:io';
-
-import 'package:flutter/services.dart';
-import 'package:logging/logging.dart';
-
-/// Syncs playback state to iOS's MPNowPlayingInfoCenter.
-///
-/// TODO: This is a workaround because audio_service doesn't set
-/// MPNowPlayingInfoCenter.playbackState on iOS (only on macOS).
-/// This causes CarPlay's Now Playing screen to not reflect the correct
-/// play/pause state when playback is started from the phone.
-/// Consider contributing a fix upstream to audio_service.
-class IosPlaybackStateSync {
-  static const _channel = MethodChannel('com.unicornsonlsd.finamp/playback_state');
-  static final _logger = Logger('IosPlaybackStateSync');
-
-  /// Sets the playback state on iOS's MPNowPlayingInfoCenter.
-  /// This is needed for CarPlay to show the correct play/pause state.
-  static Future<void> setPlaybackState({required bool isPlaying}) async {
-    if (!Platform.isIOS) return;
-
-    try {
-      await _channel.invokeMethod('setPlaybackState', {'isPlaying': isPlaying});
-      _logger.fine('Set iOS playback state to ${isPlaying ? "playing" : "paused"}');
-    } catch (e) {
-      _logger.warning('Failed to set iOS playback state: $e');
-    }
-  }
-}
